requirements.txt
langchain==0.3.7
langchain-openai==0.2.3
langchain-community==0.3.5
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0